goblin-core/requirements.txt
fastapi
uvicorn
psycopg2-binary
requests
watchdog
slpp
redis
python-dotenv
schedule
